==> gt_dtw_calculation /5_merge_sorted_files_activecasesonly.ipynb <==
import pandas as pd

tags_results = pd.read_csv('dtw_results_multi_radius_sorted.csv')
network_results = pd.read_csv('network_dtw_results_sorted.csv')
pca_results = pd.read_csv('pca_dtw_results_sorted.csv')

tags_active = tags_results[tags_results['comparison'] == 'Active Cases']
network_active = network_results[network_results['comparison'] == 'Active Cases']
pca_active = pca_results[pca_results['comparison'] == 'Active Cases']

merged_active = pd.concat([tags_active, network_active, pca_active], 
                         ignore_index=True, sort=False)

columns = merged_active.columns.tolist()
columns.remove('dtw_score')

new_columns = columns + ['dtw_score']
merged_active = merged_active[new_columns]

merged_active_sorted = merged_active.sort_values('dtw_score', ascending=True)

merged_active_sorted.to_csv('active_cases_dtw_results_sorted.csv', index=False)
